+++ education_group/api/serializers/education_group_year.py
@@ -31,8 +31,6 @@ from base.api.serializers.campus import CampusDetailSerializer
 from base.models.academic_year import AcademicYear
 from base.models.education_group_type import EducationGroupType
 from base.models.education_group_year import EducationGroupYear
-from base.models.enums import education_group_categories
-from education_group.api.serializers import utils
 from education_group.api.serializers.education_group_title import EducationGroupTitleSerializer
 from reference.models.language import Language
 
@@ -42,11 +40,8 @@ class BaseEducationGroupYearSerializer(EducationGroupTitleSerializer):
     academic_year = serializers.SlugRelatedField(slug_field='year', queryset=AcademicYear.objects.all())
     education_group_type = serializers.SlugRelatedField(
         slug_field='name',
-        queryset=EducationGroupType.objects.filter(category=education_group_categories.TRAINING),
+        queryset=EducationGroupType.objects.all()  # category=education_group_categories.TRAINING),
     )
-
-    management_entity = serializers.CharField(source='management_entity_version.acronym', read_only=True)
-    management_faculty = serializers.SerializerMethodField()
 
     # Display human readable value
     education_group_type_text = serializers.CharField(source='education_group_type.get_name_display', read_only=True)
@@ -57,14 +52,8 @@ class BaseEducationGroupYearSerializer(EducationGroupTitleSerializer):
             'code',
             'education_group_type',
             'education_group_type_text',
-            'academic_year',
-            'management_entity',
-            'management_faculty',
+            'academic_year'
         )
-
-    @staticmethod
-    def get_management_faculty(obj):
-        return utils.get_entity(obj, 'management')
 
 
 class EducationGroupYearSerializer(serializers.ModelSerializer):
